Basic functions of tetris work
The game works as it should at a basic level. There was one glitch but after many attempts at playing it did not happen again so this is not something that i will worry about at this moment, but am making a note of it. The code needs to be re commented and some variables re named for clarity. This will be done before any new changes.

Things to do:

1. Implement a 'hard drop' button. This will constantly increase the yposition until one of its blocks is touching another.

2. Improve the randomness so something like 4 pieces the same can't come up until all have, or look into a proper algorithm for it.

3. Implement a hold function. This will reset the piece imediately and replace it with a piece in the hold.

4. add graphics to the aproaching tetrimino blocks and the hold block.

5. Increase a scoring system that increases the speed each time a level is acheived.

6. Add a menu option at the begenning which can be brought up when pausing. This will include a new game and continue game options.